Use ssize_t for read return.
git-svn-id: https://svn.nmap.org/nmap-exp/devin/nse-debugger@33385 e0a8ed71-7df4-0310-8962-fdc924857419

+++ nmap_tty.h
@@ -134,6 +134,6 @@ void tty_init();
    calling method should print a status message */
 bool keyWasPressed();
 
-int tty_read(char *buffer, size_t length);
+ssize_t tty_read(char *buffer, size_t length);
 
 void set_tty_echo(int echo_on);
